feat: add debug cmds
