+++ src/constants.py
@@ -1,4 +1,5 @@
 from enum import IntEnum, auto
+import json, os
 
 # MESSAGES
 on_set_new_message = "Обновил сообщение."
@@ -98,3 +99,6 @@ RH_kick_messages = [
 ]
 
 RH_CHAT_ID = -1001147286684
+
+DEBUG = os.environ.get("DEBUG", True)
+TEAM_TELEGRAM_IDS = json.loads(os.environ.get("TEAM_TELEGRAM_IDS", "[]"))
