+++ src/handlers/__init__.py
@@ -6,4 +6,5 @@ from .help_handler import help_handler
 from .error_handler import error_handler
 
 from .admin import *
+from .debug import *
 from .group import *
