+++ src/handlers/admin/utils.py
@@ -1,6 +1,5 @@
 import json
 from typing import Iterator, Dict, List
-from functools import wraps
 
 from telegram import Bot, InlineKeyboardButton, InlineKeyboardMarkup, Update
 from telegram.ext import CallbackContext
@@ -129,27 +128,6 @@ def create_chats_list_keyboard(
         for chat in user_chats
         if authorize_user(context.bot, chat["id"], user_id)
     ]
-
-
-def admin(func):
-    """
-    A decorator to ensure that a particular function is only executed in private chats,
-    and not in group chats.
-
-    Args:
-    func (Callable): The function to be wrapped by the decorator.
-
-    Returns:
-    Callable: The wrapper function which includes the functionality for checking the chat type.
-    """
-
-    @wraps(func)
-    def wrapper(update: Update, context: CallbackContext, *args, **kwargs):
-        if update.message.chat_id < 0:
-            return  # Skip the execution of the function in case of group chat
-        return func(update, context, *args, **kwargs)
-
-    return wrapper
 
 
 def get_chat_name(bot: Bot, chat_id: int):
