+++ src/handlers/debug/__init__.py
@@ -1,0 +1,5 @@
+"""
+Module with all the telegram handlers used for debugging.
+"""
+
+from .list_jobs_handler import list_jobs_handler
